+++ lib/erl_interface/src/reg_dump.c
@@ -207,7 +207,7 @@ static int mn_send_write(int fd, erlang_pid *mnesia, const char *key, ei_reg_obj
     break;
   case EI_STR:
     if (obj->size > 0) ei_encode_string(msgbuf,&index,obj->val.s);
-    else ei_encode_long(msgbuf,&index,NULL);  /* just the NULL pointer */
+    else ei_encode_long(msgbuf,&index, (long)NULL);  /* just the NULL pointer */
     break;
   case EI_BIN:
     if (obj->size > 0) ei_encode_binary(msgbuf,&index,obj->val.p,obj->size);
